requirements.txt
pandas
tqdm
pyyaml    
openai